Make Vanta believe it's running on Ubuntu
If you try to run Vanta directly on the machine, it will show as
unsupported system (NixOS).
This tricks the agent to believe it's running in Ubuntu.
It's a safe deception as it is only informative, while the other
workarounds would mean running it inside a distrobox container - which
is less optimal.